refactor(search): convert files to typescript
