+++ src/components/search/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./search.component";
+export type { SearchProps } from "./search.component";
