--- src/components/search/search-test.stories.mdx
@@ -1,82 +1,0 @@
-import { useState } from "react";
-import { Meta, Story, Canvas } from "@storybook/addon-docs";
-import { action } from "@storybook/addon-actions";
-
-import specialCharacters from "../../__internal__/utils/argTypes/specialCharacters";
-import Search from ".";
-
-<Meta
-  title="Search/Test"
-  parameters={{
-    info: { disable: true },
-    chromatic: {
-      disable: true,
-    },
-  }}
-  argTypes={{
-    variant: {
-      options: ["default", "dark"],
-      control: {
-        type: "select",
-      },
-    },
-    placeholderSpecialCharacters: specialCharacters,
-  }}
-/>
-
-export const SearchStory = ({
-  placeholder,
-  placeholderSpecialCharacters,
-  ...args
-}) => {
-  const [value, setValue] = useState("");
-  const handleChange = (ev) => {
-    setValue(ev.target.value);
-    action("change")(ev);
-  };
-  const handleBlur = (ev) => {
-    action("blur")(ev);
-  };
-  const handleFocus = (ev) => {
-    action("focus")(ev);
-  };
-  const handleClick = (ev) => {
-    action("click")(ev);
-  };
-  const handleKeyDown = (ev) => {
-    action("keydown")(ev);
-  };
-  return (
-    <Search
-      onChange={handleChange}
-      onBlur={handleBlur}
-      onClick={handleClick}
-      onFocus={handleFocus}
-      onKeyDown={handleKeyDown}
-      value={value}
-      placeholder={placeholder || placeholderSpecialCharacters}
-      name="search_name"
-      id="search_id"
-      {...args}
-    />
-  );
-};
-
-# Alert
-
-### Default
-
-<Canvas>
-  <Story
-    name="default"
-    args={{
-      placeholder: "Search...",
-      threshold: 3,
-      searchButton: true,
-      searchWidth: "",
-      variant: undefined,
-    }}
-  >
-    {SearchStory.bind({})}
-  </Story>
-</Canvas>
